Added Index meshes to Project detail and trend-through-time button
Index meshs is another dimention and filter that includes only those meshes that have been
used in index programs for most of the time series.  Buttons allow catches to
include/exclude novel or uncommon effots.  Added a button on the project detail
page that will load the trend-through-time data for projects with matching
project codes.