Added model and view to view gear types
Added tables to hold fishnet-II data in FN013 and FN014 tables.  Also
created view and templates to show a list of all gears and gear details
page (which show information on when and where each gear was
used).  A gear tab has been added to the project detail page showing
which gears were used in a particular project and if additional FN013
and FN014 data is available.  Not tests, most (maybe all) of these
changes are temporary and were only created to scrub gear codes.