Updates to api to work with FishNet-React
This commit includes a fairly large number of changes that were
associated with the development of FishNet-React, the
client/api-consumer application.  Most of the changes are minor in
nature (addition of slug field to all serializers, creates of
Retrieve-update-destroy endpoints), and the temporary creation of a
species list endpoint and associated code (this will be removed when
we integrate uglmu-common).  We also added context to each of the
ListCreateAPI's so that the parent objects are available when new
instances are created.  Finally, django-cors-headers was added to the
application to allow api requests from a react application.