Updated script to append data from master database
This commit includes a couple of script to create/append protocol data
and update/replace existing data from the nearshore master. The
update script is very similar to the original, but provides finer
grained control over data sources and years.  This commit also
includes updates to requirements to include pyodbc and xlrd, both of
which are need to extract the data from their source files.