Removed django-widget-tweaks
This package was only being used to add a single class to the input
elements in a single form. Seems unnecessary.  Removed the package,
added the class in forms.py and everything appears as it did
before. Also removed the dependence on a non-existent migration in
the common app in the initial migration of fn_portal. All existing
tests continue to pass.