Updates to FN121 crud api and api urls and swagger documentation
This commit includes some fairly substantial changes to the crud api,
including complete test coverage for the FN121 endpoint. It also
includes a re-organization of read-only endpoints, and includes filter
classes for the FN121 and FN125 endpoints.  This commit also includes
update to the swagger api documentation which has changed recently.
All of the existing tests continue to pass.