Preparing to bundle release 0.0.2a
Not really ready for production, but we want to make the api available
in its current state for other team members to interact with.  Most of
the read-only endpoints work as expected and have good test coverage.
CRUD endpoints are still under development.  This commit also
includes new requirements for drf-yasg.  All of the existing tests
pass (with the exception of one expected to fail if a project matching
the project code cannot be found).