Fixed bug in FN125 filter and added tests
This commit fixes a bug that was caused by duplicate filter definitions
in the Fn125 table.  The filter definitions have been updated, and a
series of parameterized tests written to ensure that they work as
expected. The tests cover the FN125 attributes of the FN125 endpoint as
well as the tags, lamprey, FN126 and FN127 endpoints. All of the
existing tests continue to pass.