Final updates to FN0 design table
This is a pretty big commit and includes a number of relatively small
changes to the FN0 design tables, but also includes required updates to
the factories and associated tests.  Of particular Gear is now accessibe
through mode - it will be removed from FN121 shortly, and a constraint
has been applied to FN022 table to ensure that the seasons do not
overlap and occur within the start and end date of the project they are
associated with.  All exsisting tests continue to pass.