Removed sqlalchemy dependency and re-factored to reduce loc
This commit removes any reference to sqlalchemy - for this purpose it
increases the number of dependencies without improving
performance. This commit also includes a re-factor of some of the
functions to more generic equivalents. Additionally the number of
imports is reduced by leveraging name spaces. Finally, the whole data
upload process is wrapped in a transaction block that reverts the
updates if an error occurs.