Functional data upload
Minimal working data upload functionality. If a template database is
uploaded using the provided form, and no database errors are
triggered, the data will be inserted into fn_portal.  If an error
occurs, the insert is rolled back and an error message is displayed.
It works, but does take some time. All existing tests continue to pass.