Data upload validation with pydantic models
This is a big commit - includes pydantic models and associated tests
from the pydantic playground. Pydantic models are used to validate
data in the data upload and ensure that the data can be successfully
converted to django objects.  Errors reported by pydantic validation
are reported on the form used to upload data.  All 1100+ tests
continue to pass.