Updated upload view to use form and added logging
This commit updates the upload template view to use a django
form (with some validation) rather than getting the file from the
request object.  It also includes logging in the data
uploading/processing module and puts the odbc connection in a try
block that always closes the connection. Apparently a pyodbc context
provider does not close itself is an error.  All previously passing
tests continue to pass.