Added check for single lake and added login_required to view
This commit adds a check to verify that only a single lake is
included in the data upload. If not, an error message is
presented. This commit also adds login_required to the upload url and
present the url in the navbar only if the user is logged in. All
previously passing tests continue to pass, although the most recent
additions should have tests written for them.