Added api end point to return lake extent
This commit includes api endpoint to return all of the lakes (lake name,
abbreviation, and extent), including view, serializer, and test.  All
existing tests continue to pass.  This commit also includes some
changes to the formatting of doc strings so that they would render
properly in the django admin panel.