Initial Project Wizard API endpoint - view, serializer and test
This commit implements a basic create endpoint to consume data created
by the the FN Project Wizard. It includes some migratrations to add a
ProjectGearProcessType table and update ddlat and ddlon to dd_lat and
dd_lon.  It also includes a ProjectWizard serializer,  view to process
the parts of the submitted data, and a test to ensure that it is able
to successfully process valid data.  All previoulsy passing tests
continue to pass.