Updates to Project detail page to accommodate no samples.
There were several places on the project detail that were breaking for
new project that do not have samples yet. This commit adds some
conditional logic to the project detail template to render an
appropriate place holder if samples, catch counts of graphs are not
currently available.  It also updates the JavaScript to handle missing
data elements before operating on undefined objects.