Dropped lamijc and renamed fish_lam_id to lamid, improved validation
This commit makes a number of small changes to the validation to
include missing values (T (tag) is valid for first two characters of AGEMT,
50 is a valid gonad code). It also include migration to drop lamijc
from fn125Lamprey table - that information is perfectly duplicated by
lamijc_size and lamijc_type. Views, serializers and validation was
updated accordingly.  All of the tests continue to pass.