Updates associated with release 0.0.5a
This commit includes the creation of a FN013 api endpoint, changes
fish_lam_id to lamid, changes the size of gr code from 4 to 5, adds
notification to the upload template that you must use template v2, and
improves the admin interface for tables associated with Gear.  it also
updates the FN125 api endpoint to return the preferred age for each
fish, if one exists.