Changed gear to gr in api endpoint and added biocnt to fn123.save()
This commit updates the FN123 model to automatically calculate and
update it's biocnt value when it is saved.  Tests were added to ensure
this functionality works as expected.  Some tests had to me modified
to accommodate this new functionality. This commit also updates the
gear_effort_process type endpoint so that the gear column is called
'gr' rather than 'gear' to ensure consistency with other columns that
reference gear codes.  All previously passing tests continue to pass.