Fixed Xslime schema and added biocnt default to detail page
This commit includes a small update the FN121 schema - it was using
the wrong enumeration field for xslime. This has been fixed, and a
couple of additional tests added.  A default value of 0 was added the
project summary page for biocnts - if there are not biocnts associated
with the species, the total should be 0, not 'None'. All tests
continue to pass.