Refined validators for fn121.grdepmin and fn125.rwt
This commit updated the validator for fn121.grdepmin to allow 0
values, and ensure that it always less than grdepmax if both are
populated. fn125.rwt was also updated to a float to allow very small
fish to be measured without loosing precision or rounding to an
integer which often caused other validation errors.  All previously
passing tests continue to pass.