Added FN124 to upload process and bumped to version 0.0.6
This commit adds a FN124 pydantic schema and tests, and activated the
code to fetch, process, and insert FN124 data from the template
database. It also updated the upload view to redirect to the project
detail if there is just project code, or add a query filter to the
project list if multiple project codes are included in the upload.
All of the existing test pass. This commit will be tagged with version
0.0.6.