Added filters for FN124 endpoint, filter for sam every where
This commit adds filters for our fn124 api endpoint. It has the same
parent filters as FN125, as well as filters for siz and sizcnt. This
commit also adds filters for sam to every endpoint, not sure how that
was missed earlier, and adds a couple of additional files that should
have been committed when the FN124 upload and validation functionality
was added.  All existing tests continue to pass.