Added geom to FN121 and spatial filters to all child tables
This commit add a geometry field to the FN121 table and updates it
using the available dd_lat and dd_lon.  It also add spatial filterset
class and adds spatial filters for FN121 and all child
tables. Filters are provided for both polygons and points.  All
existing tests continue to pass, although tests for the spatial
filters have not been added and need to be.