Work-around for buffered point transform issues
There seems to be a bug in the django geos library that does not
properly back trasform the polygon produced from a buffered
point. This commit includes a work around for that limitation and
includes a function that does the transformation in postgis.  All
previously passing tests continue to pass.