Updates to gr and grtp filters, added TISSUE codes, and updated AGEST
This commit fixes the filters for api endpoints from FN122 and
below. The original ones were not updated to accommodate the FN028
table.  A validation was added for TISSUE codes and the valid codes
for AGEST were tweaked to remove 'other' and add vertebrae. All
previous tests continue to pass.