Added Pydantic schema and tests for FN012 objects
this commit includes a Pydantic model and two tests files for FN012
objects.  One test file verifies that the regular expressions used by
the pydantic model work as expected, while the other test ensures the
the validator as a whole works as expected and returns meaning error
messages if invalid data is submitted.  All previously passing tests
contiue to pass.