Added FN012Project List and FN012 to project detail
This commit add templates to list protocols and the default FN012
values for each.  It also updated the project detail page to include
the FN012 values used in that project.  It also updated the api
endpoint for FN012Protocol to include an option to download values to
csv to facilitate populating the template databases.  All previously
passing tests continue to pass.