Tweaked the FN012 bounds and updated pydantic schema
This commit updates the bounds on a number of fields in the FN012
table and pydantic schema. Specifically it broadens the bounds on the
condition factor and increased the max_rwt value. It also adds some
conditional logic to only require sizatt and sizint when sizsam is 2
or 3 (when fn124 data will be collected).  Additional tests were added
to ensure the new functionality works as expected.  All previously
passing tests continue to pass.