Updated to FN012 model to allow agedec='00'
The last commit added '00' as a valid agedec value in the schema, but
there were several model methods that were not recognizing '00' as a
valid. This commit updates those methods and adds unit tests to
ensure that fn012 objects can be created with agedec='00'.  All
previously passing tests continue to pass.