Updates FN012, FN124, FN126 and FN127 validators
This commit implements a number of small changes to the validators
that had been suggested by the GLIS data in team. Specifically, the
biological limits fields are no longer required if Biosam=0, the
minimum a the siz in FN124 is now 10, and empty strings are
automatically replaced with null values for optional enumeration
fields in FN126 and FN127 tables.  Additionally a '---' is displayed
in the Lake/Protocol table rather than 'None' when no value is
available.  All tests continue to pass.