Upgraded to python 3.9 and Django==3.2.12
This commit includes a number of small updates associated with
upgrading from python 3.7 to 3.9 and Django 2.2 to 3.2.  Most of
changes are associated with the addition of an id auto-field to the
models. Some other small changes where made to address warnings raised
at the end of the test runs.  This commit also updates all of the
requirements and their dependencies to current stable version. All
tests continue to pass without issuing any warnings.