Re-factored data upload code into helper functions and added tests
This commit re-factors much of the code associated with recreating
records in the database from pydantic data models. New records are
created, missing records are deleted, and existing records that have
changed are updated accordingly. Unit tests are also included to
ensure that records are updated as expected. All previously passing
tests continue to pass.