Updated bounds on dd_lon, validation error msg, project detail spc
This commit implements a number of small changes to the templates and
eastern most bounds of validation for dd_lon. The project detail page
was updated to correctly show the species scientific name is the
common name was unavailable.  The validation error message now
correctly uses 'were' rather than 'where', and the eastern most of
dd_lon for validation was increased to -74.32 to accommodate the
St. Lawrence River. All previously passing tests continue to pass.