Updates to FN012Protocol views, template, constraints, and admin
This commit includes a number of small changes associated with
FN012Protocols.  The csv upload in the admin was improved to
automatically format species codes and ensure that lake, protocol, and
sizatt are all uppercase. It also include a pydantic model for
FN012Protocol objects, including tests, although it is not currently
used.  This commit also tweaks the validation constrains for a couple
of values to reflect data actually collected in some projects for some
species.  Finally, the Protocol list and view were updated to show if
the protocol has been confirmed and should be considered active.  All
previously passing tests continue to pass.