Added api listview, url, filters, and serialize for FN121Limno
This commit adds the basic elements of the api endpoint for FN121
Limno data.  It includes a read-only list view, a filter set that
accepts all of the FN121 filters plus those associated with
limnological parameters and a read only serializer that includes
prj_cd and sam. Not tests have been written for these components, but
all existing tests continue to pass.