Update size serializer for some endpoints, added indexes to FN127
This commit implements a number of small fixes that are intended to
improve the api's performance or usability. The FN122 endpoint now
uses a serializer that is explicitly read-only, and several of the
endpoint have now been bumped up to return large page sizes by
default. The commit also includes some migration that place indexes on
the FN125 and FN127 tables to make returning the default age for each
FN125 record more efficient.  Finally, the pydandic schema for FN125
objects was tweaked to accommodate even skinnier fish in Lake Erie.
All test continue to pass.