Added FN026Subspaces between FN026 and FN121, including migrations
This commit includes all of the migration files associated with the
insertion of FN026Subspace table.  It assumes that each subspace is
equal to each space (which is currently true).  The commit also
includes basic tests for FN026Subspace model attributes, and updates
to all existing tests to ensure that they continue to pass.