Renamed fields in FN121 to match field names in FN121
There are several fields that were renamed to include a 0 suffix to
follow the FN convention for first and second readings. This was
implied, but is now explicit.  Additionally, the 'x' as dropped from
slime field.  All of the fixtures, factories, tests, views, and
templates were updated to use the new field names. All previously
passing tests continue to pass.