New Models for template4. FN121Weather tests.
This commit includes the models and migrations for the new tables
included in tempalte4.  It also includes a factory and a series of
unit tests for the FN121Weather model (which contains considerably
more logic than you might think).  One of the serializers that fetches
management units needed to be updated to account for the new model
structure associated with management units.  All previously passing
tests continue to pass.