Tweaks to model and api endpoint to match template 4
This commit includes a number of small additions to the models or api
views and serializers to match values added in template4.  For the
most part, these are additional fields that are rarely used, but are
included in the current template.  This commit also includes a
migration file to add salmonid stream and lake ages, as well as
associated changes to data upload utilities.  All previously passing
tests continue to pass.