Changed pydandic classes to factories for dynamic choices
This is a big commit and updates several of the pydanic validation
models to factory functions that accept a list of choice that can be
used to instantiate pydantic classes. This allows them to be
instantiate with choice from the database which can be updated from
django admin dashboard. the choices are currently hard-coded lists or
dictionaries, but these will be replaced shortly.  Several of the
associated tests and the project upload functionality had to be
updated accordingly.  All of the existing tests continue to pass.