View, filters, serializer and tests for FN121Electrofishing
This commit adds api components - a list view, filters, serializers,
and tests for FN121Electofishing items. It also incorporates them into
the data upload process.  This commit also includes a migration that
changes the names from _minimum and _maximum to _min and _max to match
the conventions used in the template database. All previously passing
tests continue to pass.