Renamed FN122Transects to FN121GpsTracks
This is a fairly large commit that includes a large number of small
updates - essentially every reference to FN122 Transect has been
replaced with a reference to FN121 Gps Track.  This commit also
includes a migration that removes the old table and creates a new
one. All previously passing tests continue to pass.